requirements.txt
Flask==3.1.1
uwsgi==2.0.30
requests==2.32.4
ollama==0.5.3
pyyaml==6.0.2
-e .